refactor(platform-browser-dynamic): relocate `DOMTestComponentRenderer` to `@angular/platform-browser`
This commit moves `DOMTestComponentRenderer` to `@angular/platform-browser/testing`, allowing the Angular CLI to eliminate its dependency on `@angular/platform-browser-dynamic`, which would no longer be required for new projects.
