+++ packages/platform-browser-dynamic/testing/src/private_export_testing.ts
@@ -6,5 +6,5 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-export {DOMTestComponentRenderer as ɵDOMTestComponentRenderer} from './dom_test_component_renderer';
 export {platformCoreDynamicTesting as ɵplatformCoreDynamicTesting} from './platform_core_dynamic_testing';
+export {ɵDOMTestComponentRenderer} from '@angular/platform-browser/testing';
