+++ packages/platform-browser-dynamic/testing/src/testing.ts
@@ -7,11 +7,9 @@
  */
 
 import {createPlatformFactory, NgModule, PlatformRef, StaticProvider} from '@angular/core';
-import {TestComponentRenderer} from '@angular/core/testing';
 import {ɵINTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS as INTERNAL_BROWSER_DYNAMIC_PLATFORM_PROVIDERS} from '@angular/platform-browser-dynamic';
 import {BrowserTestingModule} from '@angular/platform-browser/testing';
 
-import {DOMTestComponentRenderer} from './dom_test_component_renderer';
 import {platformCoreDynamicTesting} from './platform_core_dynamic_testing';
 
 export * from './private_export_testing';
@@ -32,6 +30,5 @@ export const platformBrowserDynamicTesting = createPlatformFactory(
  */
 @NgModule({
   exports: [BrowserTestingModule],
-  providers: [{provide: TestComponentRenderer, useClass: DOMTestComponentRenderer}],
 })
 export class BrowserDynamicTestingModule {}
