+++ packages/platform-browser/testing/src/dom_test_component_renderer.ts
