+++ packages/platform-browser/testing/src/private_export_testing.ts
@@ -6,10 +6,4 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-/**
- * @module
- * @description
- * Entry point for all public APIs of the platform-browser/testing package.
- */
-export * from './browser';
-export * from './private_export_testing';
+export {DOMTestComponentRenderer as ɵDOMTestComponentRenderer} from './dom_test_component_renderer';
